requirements.txt
requests==2.31.0
supabase==2.7.4
httpx==0.27.2
python-dotenv==1.0.0
playwright==1.44.0
